Added additional repository methods
`IsTemplateRepository`: need to check of a repo is a template
repository, else it cannot be used for the new repo

`GetTemplateRepositoryOwner`: If the template repo owner is not the same
as the destination for the new repo, there should be an `owner` field in
the `Repository` struct.  Check this and pass it to the lower repo
create method.